Update README.md
switch and if-else-if statements

Rewrite the program using if and else if statements (multi-way else-if) rather than a switch statement.

Use a trailing else in your new version corresponding to the default case in the switch statement of the original program.

Reminder: The char type is used for variables that hold a single character. When using character literals in a program (like 'A') the character must be enclosed in single quotes, not double quotes that are used with string literals.